requirements.txt
streamlit==1.44.1
openai==1.14.3
pandas>=1.5
rapidfuzz
PyPDF2==3.0.1
python-docx==0.8.11
requests>=2.31.0
reportlab==3.6.12
